Consolidated radar and hallsensor code, began braking class
Created the braking_inputs.py file, in which I have a Radar class and a HallSensor class, each of which is threaded such that distance to object and cart speed are constantly being updated. Created braking class in which I instantiate the classes within braking_inputs so that speed of cart and distance to object can be constantly taken into account for braking purposes. Ballout.